Move instantiation of real-time components to separate module
