+++ application/src/main/java/org/opentripplanner/standalone/configure/ConstructApplicationFactory.java
@@ -48,6 +48,7 @@ import org.opentripplanner.transit.service.TimetableRepository;
 import org.opentripplanner.transit.service.TransitService;
 import org.opentripplanner.updater.trip.TimetableSnapshotManager;
 import org.opentripplanner.updater.trip.TripPatternCache;
+import org.opentripplanner.updater.trip.configure.TripUpdateModule;
 import org.opentripplanner.visualizer.GraphVisualizer;
 
 /**
@@ -70,6 +71,7 @@ import org.opentripplanner.visualizer.GraphVisualizer;
     StopConsolidationServiceModule.class,
     StreetLimitationParametersServiceModule.class,
     TransitModule.class,
+    TripUpdateModule.class,
     VehicleParkingServiceModule.class,
     VehicleRentalRepositoryModule.class,
     VehicleRentalServiceModule.class,
